Changed datatype of the variable hoursStreamed to double

+++ src/main/java/edu/kea/kab/model/Consumption.java
@@ -15,6 +15,6 @@ public class Consumption {
     @Enumerated(EnumType.STRING)
     private Categories category;
 
-    private Long hoursStreamed;
+    private double hoursStreamed;
 
 }
